backend/requirements.txt
amqp==5.3.1
asgiref==3.11.0
async-timeout==5.0.1
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.10
django-filter==25.2
djangorestframework==3.16.1
exceptiongroup==1.3.1
idna==3.11
kombu==5.6.2
newsapi-python==0.2.7
packaging==26.0
prompt_toolkit==3.0.52
psycopg==3.3.2
psycopg-binary==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.1.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.5.0
